1.18
Updated to 1.18
Note that if upgrading a world from a previous Minecraft version, fans will need to be re-activated to work!
