+++ gradle.properties
@@ -3,24 +3,24 @@
 
 # Fabric Properties
 	# check these on https://modmuss50.me/fabric.html
-	minecraft_version=1.17.1
-	yarn_mappings=61
+	minecraft_version=1.18-rc3
+	yarn_mappings=2
 	tiny_version=2
-	loader_version=0.11.7
+	loader_version=0.12.5
 
 # Mod Properties
 	mod_version = 1.3.3
 	maven_group = com.shnupbups
 	mod_name = trapexpansion
-	version_meta = fabric-mc1.17
+	version_meta = fabric-mc1.18
 
 # Dependencies
 	# check on https://modmuss50.me/fabric.html
-	fapi_version=0.40.1+1.17
+	fapi_version=0.43.1+1.18
 
 # Other Stuff
 	# check on maven at https://maven.shedaniel.me/me/shedaniel/RoughlyEnoughItems-fabric/
-	rei_version = 6.0.279-alpha
+	rei_version = 7.0.339
 	
-	# check on maven at https://maven.terraformersmc.com/releases/com/terraformersmc/modmenu
-	modmenu_version = 2.0.14
+	# check on maven at https://maven.terraformersmc.com/releases/com/terraformersmc/modmenu/
+	modmenu_version = 3.0.0
